Made minor improvments in Entry Conversion
By making a very slight change in the BackendEntry API, I was able to reduce a large portion of the conversion code built into the "process" function. Additionally, I reduced the Manager's prototype by seperating the "process" method from the prototype, and replacing all calls to "this.process()" with "process.call(this)".