Added a proper shim for the Console Backend & fixed RunKit
The Console Backend didn't quite feel flexable enough. After all, each
platform seems to have different capabilities and limitations associated
with the native console object, so I thought it would be a good idea to
make the Console Backend able to be shimmed, while only requiring a
minimum of just the "log" method to be available.

Oh yeah, and I also fixed the "package.json" file so it pointed to the
"runkit.js" file as the RunKit example. Kind of forgot to do that before
the last commit.