Split files & build targets
In an effort to push the project toward the 1.0.0-ALPHA.1 release, I
made a number of improvments in the project stucture including splitting
the master source file into multiple smaller files, adding a build
target to compile a CommonJS compatible version of the library, and
stablizing the RunKit sample file using the new CommonJS target. I also
took the time to upgrade all of the dev-dependencies to get rid of those
nasty security warnings.