Updated License Agreement
Upon further consideration, I have decided that it would be better to
offer Woodsman under the LGPL agreement, rather than the AGPL agreement.
I understand that this may lead to issues in cases where the end user
may choose to serve Woodsman in compiled form from a web server without
distributing the source code, but it is a risk I am willing to take in
order to promote the actual usage of the library.